generic impls of error responses
